requirements.txt
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.8.0
requests>=2.25.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.60.0
numpy>=1.20.0
pandas>=1.3.0
python-dotenv>=1.0.0
